Update main.py
I believe that this has fixed the issue with old torrents downloading every time. Rsync has a few options to prevent this but each one I found seemed to be something that needed to be patched in and caused more problems than it solved. Instead we find the newest file in the /watch directory and use that to create a file list.